updated zio program to read individual vector files

+++ benchmarks/python/preprocess.py
@@ -9,25 +9,41 @@ import sys
 from elastiknn.api import Vec
 from elastiknn.utils import ndarray_to_sparse_bool_vectors
 from imagehash import phash
+from math import sqrt
 from typing import List
 
 from PIL import Image
 from tqdm import tqdm
 
 
+def mkdirp(d):
+    if not os.path.exists(d):
+        os.mkdir(d)
+
+
+def rmrf(f):
+    if os.path.exists(f):
+        os.remove(f)
+
+
+def touch(f):
+    with open(f, "w"):
+        pass
+
+
 def annb(url: str, datadir: str):
     pass
 
-
-def amazon_reviews(url: str, datadir: str):
-    pass
 
 
 def amazon_raw(url: str, datadir: str):
     pass
 
 
-def amazon_phash(url: str, datadir: str, n: int = sys.maxsize):
+def amazon_phash(url: str, datadir: str, dim: int = 4096, n: int = sys.maxsize):
+
+    assert sqrt(dim) == int(sqrt(dim))
+    hash_size = int(sqrt(dim))
 
     if not os.path.exists(datadir):
         os.mkdir(datadir)
@@ -38,46 +54,36 @@ def amazon_phash(url: str, datadir: str, n: int = sys.maxsize):
         wget.download(url, metafile)
 
     imgsdir = f"{datadir}/images"
-    if not os.path.exists(imgsdir):
-        os.mkdir(imgsdir)
+    mkdirp(imgsdir)
 
-    # Keep a set of missing URLs.
-    missingfile = f"{imgsdir}/missing.txt"
-    if not os.path.exists(missingfile):
-        with open(missingfile, "w"):
-            pass
-    with open(missingfile) as fp:
-        missing_urls = set(fp.read().split('\n'))
-    missingfp = open(missingfile, "w")
+    vecsdir = f"{datadir}/vecs"
+    mkdirp(vecsdir)
 
-    # Download images.
     with gzip.open(metafile) as gzfp:
         lines = list(tqdm(gzfp, desc=f"Reading {metafile}"))[:n]
-        with tqdm(lines, desc="Downloading images") as pbar:
+        with tqdm(lines, desc="Processing images") as pbar:
             for d in map(eval, lines):
-                if "imUrl" in d and d["imUrl"] not in missing_urls:
-                    asin, url = d['asin'], d['imUrl']
-                    imgfile = f"{imgsdir}/{asin}.jpg"
-                    pbar.set_description(f"Processing {url} to {imgfile}")
-                    if not os.path.exists(imgfile):
-                        try:
-                            urllib.request.urlretrieve(url, imgfile)
-                        except urllib.error.HTTPError as ex:
-                            missing_urls.add(url)
-                            missingfp.write(f"{url}\n")
-                            missingfp.flush()
-                            print(f"Caught exception {ex} for url {url}", file=sys.stderr)
-                        pbar.update()
-
-    vecfile = f"{datadir}/vectors.json"
-
-    with open(vecfile, "w") as vfp:
-        for imgfile in tqdm(os.listdir(imgsdir), desc="pHashing images"):
-            if imgfile.endswith(".jpg"):
-                img = Image.open(f"{imgsdir}/{imgfile}")
-                ph = phash(img, 64)
-                for vec in ndarray_to_sparse_bool_vectors(ph.hash.reshape((1, ph.hash.size))):
-                    vfp.write(f"{json.dumps(vec.to_dict())}\n")
+                if "imUrl" not in d:
+                    continue
+                asin, url = d['asin'], d['imUrl']
+                pbar.set_description(f"Processing {url}")
+                imgfile = f"{imgsdir}/{asin}.jpg"
+                imgfile_missing = f"{imgsdir}/{asin}.missing"
+                vecfile = f"{vecsdir}/{asin}.json"
+                if not os.path.exists(imgfile) and not os.path.exists(imgfile_missing):
+                    try:
+                        urllib.request.urlretrieve(url, imgfile)
+                    except urllib.error.HTTPError as ex:
+                        print(f"Caught exception {ex} for url {url}", file=sys.stderr)
+                        rmrf(imgfile)
+                        touch(imgfile_missing)
+                if os.path.exists(imgfile) and not os.path.exists(vecfile):
+                    img = Image.open(imgfile)
+                    ph = phash(img, hash_size)
+                    for vec in ndarray_to_sparse_bool_vectors(ph.hash.reshape((1, ph.hash.size))):
+                        with open(vecfile, "w") as fp:
+                            json.dump(vec.to_dict(), fp)
+                pbar.update()
 
 
 def main(argv: List[str]) -> int:
